+++ packages/at_client_flutter/example/lib/apkam_example.dart
@@ -131,10 +131,14 @@ class _ApkamExamplePageState extends State<ApkamExamplePage> {
     try {
       _logger.info('Initiating simulated request for $appName');
 
-      final atLookup = AtLookupImpl(
-        currentAtSign,
-        atClient.getPreferences()!.rootDomain,
-        atClient.getPreferences()!.rootPort,
+      final atLookup = AtLookUp.withSecureSocket(
+        atSign: currentAtSign,
+        rootDomain: AtRootDomain(
+          atClient.getPreferences()!.rootDomain,
+          atClient.getPreferences()!.rootPort,
+        ),
+        secureSocketConfig: SecureSocketConfig(),
+        authenticator: null,
       );
 
       // STEP 1: Generate an OTP using the MANAGER client